Rename student to member
